# new_project/contracts/openapi/investment-api.yaml
openapi: 3.1.0
info:
  title: Personal Investment API
  version: 0.2.0
  description: 本地开发切片。所有金额以原币种两位小数最小单位的十进制字符串传输，绝不使用浮点数。
servers:
  - url: http://127.0.0.1:<API_PORT>
paths:
...
  /api/v1/ledger/imports:
    post:
      summary: 创建历史账本 dry-run 导入作业
      description: 仅接受已扫描的私有 IMPORT 文件。服务端使用映射生成不可变预览；不会在此步骤写入账本。重提同一文件会使其未确认旧预览失效。
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyImportRequest'
      responses:
        '202':
          description: 已创建可轮询的本地 dry-run 作业。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyImportJob'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/jobs/{jobId}:
    get:
      summary: 查询历史导入 dry-run 作业
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      responses:
        '200':
          description: 导入作业及逐行预览结果。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyImportJob'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /api/v1/ledger/imports/{jobId}/confirm:
    post:
      summary: 以 checksum 确认历史导入并原子追加账本事实
      description: 仅 SUCCEEDED、未过期且 checksum 完全匹配的预览可确认。确认后所有事实 sourceType=IMPORT 且关联 importExportFileId；旧 Dashboard ic/roll 仅导入其费用并标记 ROLL_FEE_ONLY。
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyImportConfirm'
      responses:
        '200':
          description: 已原子确认。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyImportJob'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /api/v1/portfolio/manual-valuations:
    post:
      summary: 追加手工估值事实
      description: 仅影响组合读模型，不修改账本分录、成本或持仓批次。仅支持股票、ETF、期权；币种必须等于标的原生币种，单价和用户级总市值二选一。
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualValuationCreate'
      responses:
        '201':
          description: 已追加不可变手工估值事实
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualValuation'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    Problem:
      description: RFC 7807 风格错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    LegacyImportRequest:
      type: object
      required: [importExportFileId, format]
      properties:
        importExportFileId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        format:
          type: string
          enum: [LEGACY_DASHBOARD_JSON, LEGACY_SQLITE]
        snapshotId:
          type: [string, 'null']
          description: LEGACY_SQLITE 必填的 snapshots.id；JSON 不得提供。
          maxLength: 64
        currencyMappings:
          type: array
          items:
            $ref: '#/components/schemas/LegacyCurrencyMapping'
        instrumentMappings:
          type: array
          items:
            $ref: '#/components/schemas/LegacyInstrumentMapping'
        dividendEntitlementOverrides:
          type: array
          items:
            $ref: '#/components/schemas/LegacyDividendEntitlementOverride'
        optionExpiryAttestations:
          type: array
          items:
            $ref: '#/components/schemas/LegacyOptionExpiryAttestation'
    LegacyCurrencyMapping:
      type: object
      required: [module, currency, amountUnit, cashAccountId]
      properties:
        module:
          type: string
        action:
          type: [string, 'null']
          description: 为空时匹配该 module 的所有 action；同 module/action 不得重复。
        currency:
          type: string
          enum: [CNY, USD]
        amountUnit:
          type: string
          description: 适用于 amount、margin 与 fee；price 始终按原币种小数精确乘以 100。
          enum: [LEGACY_CNY_WAN, ORIGINAL_CURRENCY_DECIMAL]
        cashAccountId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    LegacyInstrumentMapping:
      type: object
      required: [module, symbol, instrumentId]
      properties:
        module:
          type: string
        symbol:
          type: string
        instrumentId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    LegacyDividendEntitlementOverride:
      type: object
      required: [sourceRow, entitlementDate]
      properties:
        sourceRow:
          type: integer
          minimum: 1
        entitlementDate:
          type: string
          format: date
    LegacyOptionExpiryAttestation:
      type: object
      required: [sourceRow, expiryOutcome]
      properties:
        sourceRow:
          type: integer
          minimum: 1
        expiryOutcome:
          type: string
          enum: [WORTHLESS]
    LegacyImportConfirm:
      type: object
      required: [expectedChecksum]
      properties:
        expectedChecksum:
          type: string
          pattern: '^[a-f0-9]{64}$'
    LegacyImportJob:
      type: object
      required: [jobId, importPreviewId, importExportFileId, format, previewChecksum, status, expiresAt,
                 applicableCount, needsReviewCount, lines]
      properties:
        jobId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        importPreviewId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        importExportFileId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        format:
          type: string
          enum: [LEGACY_DASHBOARD_JSON, LEGACY_SQLITE]
        sourceSnapshotId:
          type: [string, 'null']
        previewChecksum:
          type: string
          pattern: '^[a-f0-9]{64}$'
        status:
          type: string
          enum: [SUCCEEDED, FAILED, NEEDS_REVIEW, COMMITTED, EXPIRED]
        expiresAt:
          type: string
          format: date-time
        applicableCount:
          type: integer
          minimum: 0
        needsReviewCount:
          type: integer
          minimum: 0
        lines:
          type: array
          items:
            $ref: '#/components/schemas/LegacyImportPreviewLine'
    LegacyImportPreviewLine:
      type: object
      required: [sourceRow, status]
      properties:
        sourceRow:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [APPLICABLE, NEEDS_REVIEW, SKIPPED, FAILED]
        code:
          type: [string, 'null']
        operation:
          type: [string, 'null']
          enum: [EXTERNAL_FUNDING, EXTERNAL_WITHDRAWAL, SPOT_BUY, SPOT_SELL, DIVIDEND, INTEREST, OPTION_OPEN,
                 OPTION_CLOSE, OPTION_EXPIRE_ALL, FUTURES_MARGIN_IN, FUTURES_OPEN, FUTURES_CLOSE, FEE, null]
        occurredOn:
          type: [string, 'null']
          format: date
        cashAccountId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        instrumentId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        quantity:
          type: [string, 'null']
        unitPriceCent:
          type: [string, 'null']
          pattern: '^[1-9][0-9]*$'
        pricePoints:
          type: [string, 'null']
        amountCent:
          type: [string, 'null']
          pattern: '^(?:0|[1-9][0-9]*)$'
        initialMarginCent:
          type: [string, 'null']
          pattern: '^[1-9][0-9]*$'
        feeCent:
          type: [string, 'null']
          pattern: '^(?:0|[1-9][0-9]*)$'
        entitlementDate:
          type: [string, 'null']
          format: date
        note:
          type: [string, 'null']
          maxLength: 1000
...
    ManualValuationCreate:
      type: object
      required: [instrumentId, valuationDate, currency]
      properties:
        instrumentId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        valuationDate:
          type: string
          format: date
        currency:
          type: string
          enum: [CNY, USD]
        unitPriceCent:
          type: [string, 'null']
          description: 原币种最小单位单价；与 marketValueCent 二选一。
          pattern: '^[1-9][0-9]*$'
        marketValueCent:
          type: [string, 'null']
          description: “用户 + 标的”的原币种最小单位总市值，只在组合总览中计一次；与 unitPriceCent 二选一。
          pattern: '^[1-9][0-9]*$'
        validUntil:
          type: [string, 'null']
          format: date-time
        note:
          type: [string, 'null']
          maxLength: 1000
    ManualValuation:
      allOf:
        - $ref: '#/components/schemas/ManualValuationCreate'
        - type: object
          required: [manualValuationId, priority, valuationStatus]
          properties:
            manualValuationId:
              type: string
              pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
            priority:
              type: integer
              enum: [100]
            valuationStatus:
              type: string
              enum: [ACTIVE, EXPIRED]
...
    Problem:
      type: object
      required: [code, message, traceId, details]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          type: array
          items:
            type: string